requirements.txt
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
llama-cpp-python>=0.2.0
huggingface-hub>=0.19.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=2.0.0
mcp>=0.9.0
sentencepiece>=0.1.99
protobuf>=3.20.0